Added foot locking method selector with just the current method
